fix: resolve critical review issues from parallel code review
- C1+C2: fix posted_ts=0 bug — fetch metadata before timestamp filtering,
  use real 'posted' from API instead of parser placeholder
- C3: add retry_failed_eh_downloads() — failed downloads are re-queued
  up to max_retry_count instead of being permanently abandoned
- C4: pending_queue pop is deferred (download processor handles lifecycle)
- I3: remove unused _notifier/_telegraph params from EhEngine::new
- I4: wire EhentaiConfig.max_retry_count into EhDownloadProcessor
- I7: fix handle_eunsub to search subscriptions by query prefix instead
  of reconstructing task value without filters
- I9: use spawn_blocking for ZIP extraction in process_telegraph
- M2: fix EhFilter::matches to use float comparison instead of u8 truncation
- Fix 48h scan test to verify working behavior instead of documenting bug
- Rate limit default: 7GB per 7 days (168h)

+++ src/db/types/eh_filter.rs
@@ -59,7 +59,7 @@ impl EhFilter {
     /// Check if a gallery matches all filter criteria.
     pub fn matches(&self, gallery: &EhGallery) -> bool {
         if let Some(min_rating) = self.min_rating {
-            if (gallery.rating as u8) < min_rating {
+            if gallery.rating < min_rating as f64 {
                 return false;
             }
         }
